Create 67.AddBinary.cpp
First we make copies of the string, pad the smaller string to match length. Then we iterate from the last element and use conditionals to check if they add up to 1 or 0, and if there is a carry, we use a carry boolean to carry it over. Lastly, we reverse the string to get the final answer.